Consider sidebar for ordering of chapters/files
GitHub offers the option to define a custom sidebar with custom ordering of markdown files instead of just using alphabetical order. If the user chooses to create this custom sidebar, a sidebar-file is created in which he can define the ordering. The "pages" tab usually displayed to the right of a github-wiki then is collapsed by default and the sidebar gets shown instead.
For wikidoc this means that if a sidebar-file exists, the ordering can be taken from there.
Example usage: https://github.com/KernelHaven/KernelHaven/wiki